Use async_manual_reset_event instead of timer
The benchmark hangs when run with 1 million tasks. I haven't looked into it, but perhaps there is an inefficient implementation of the timer in unifex.

As a POC, I reimplemented the unifex benchmark to spawn N tasks and have them wait until all N are created before completing the program. I did not change the other benchmarks.